upd: move some methods from class Check to new class Action upd EditUserTests, tryChangeEmailToExists: move deleting 2nd user to After annotation
